requirements.txt
alembic==1.10.3
anyio==3.6.2
attrs==22.2.0
certifi==2022.12.7
charset-normalizer==3.1.0
click==8.1.3
coverage==7.2.3
exceptiongroup==1.1.1
factory-boy==3.2.1
Faker==18.4.0
fastapi==0.95.0
greenlet==2.0.2
gunicorn==20.1.0
h11==0.14.0
idna==3.4
importlib-metadata==6.1.0
importlib-resources==5.12.0
inflection==0.5.1
iniconfig==2.0.0
Mako==1.2.4
MarkupSafe==2.1.2
packaging==23.0
pluggy==1.0.0
pydantic==1.10.7
pytest==7.2.2
pytest-cov==4.0.0
pytest-factoryboy==2.5.1
python-dateutil==2.8.2
python-dotenv==1.0.0
requests==2.28.2
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.9
starlette==0.26.1
tomli==2.0.1
typing-extensions==4.5.0
urllib3==1.26.15
uvicorn==0.21.1
zipp==3.15.0
